UJS-453: Fixed stone weight.

See merge request rubber-duck/uvel!555

+++ database/migrations/2018_01_09_075251_create_stones_table.php
@@ -17,7 +17,7 @@ class CreateStonesTable extends Migration
             $table->increments('id');
             $table->string('nomenclature_id');
             $table->integer('type');
-            $table->float('weight');
+            $table->decimal('weight', 10, 3);
             $table->float('carat');
             $table->unsignedInteger('size_id');
             $table->unsignedInteger('style_id');
