+++ database/migrations/2018_01_15_094831_create_product_stones_table.php
@@ -19,7 +19,7 @@ class CreateProductStonesTable extends Migration
             $table->integer('model_id')->unsigned();
             $table->integer('stone_id')->unsigned();
             $table->integer('amount');
-            $table->integer('weight');
+            $table->decimal('weight', 10, 3);  
             $table->enum('flow', ['yes', 'no'])->default('no');
             $table->timestamps();
             $table->softDeletes();
